Firebase helper functions + First time login route
When the user logs into the system for the first time, they get redirected to the preference page. If they are a returning user, It directs them straight to the camera page.

Added helper firestore functions to firebase Js to reduce repeated functions. updateDocument is the only tested one so far